feat: enforce rules
disallow rent if no available motorcycles
disallow delete motorcycle with rent active
disallow drivers without category A from renting

refers to KAN-18

+++ RentApp.ApiService/Server/MotorcycleEndpoints.cs
@@ -104,6 +104,20 @@ public static class MotorcycleEndpoints
             {
                 return result;
             }
+            motorcycleApiDataModel.LicensePlate = oldId;
+            motorcycleApiDataModel.ActiveUserlId = null;
+            bool rentalCleared = false;
+            await client.PutMotorcycle(motorcycleApiDataModel, () =>
+            {
+                rentalCleared = true;
+            }, (s) =>
+            {
+                result = TypedResults.Problem(detail: s);
+            });
+            if (!rentalCleared)
+            {
+                return result;
+            }
             await client.DeleteMotorcycle(oldId, () =>
             {
                 result = TypedResults.Ok();
@@ -128,7 +142,7 @@ public static class MotorcycleEndpoints
                     result = TypedResults.Ok();
                 }, (s) =>
                 {
-                    result = TypedResults.Problem(detail: s);
+                    result = TypedResults.BadRequest(s);
                 });
             }
             return result;
